+++ src/sys/mem/heap.rs
@@ -1,22 +1,23 @@
+use crate::sys::x86::addr::VirtAddr;
+
 use linked_list_allocator::LockedHeap;
 use x86_64::structures::paging::{FrameAllocator, Mapper, Page, PageTableFlags};
-use x86_64::VirtAddr;
 
 #[cfg_attr(not(feature = "userspace"), global_allocator)]
 static ALLOCATOR: LockedHeap = LockedHeap::empty();
 
-pub const HEAP_START: u64 = 0x4444_4444_0000;
+pub const HEAP_START: usize = 0x4444_4444_0000;
 
 #[cfg(target_arch = "x86_64")] // TODO: Remove
 pub fn init_heap() -> Result<(), ()> {
     let mapper = super::mapper();
 
     // Use half of the memory for the heap
-    let heap_size = (super::memory_size() / 2) as u64;
+    let heap_size = super::memory_size() / 2;
     let heap_start = VirtAddr::new(HEAP_START);
 
     let pages = {
-        let heap_end = heap_start + heap_size - 1u64;
+        let heap_end = heap_start + heap_size - 1;
         let heap_start_page = Page::containing_address(heap_start);
         let heap_end_page = Page::containing_address(heap_end);
         Page::range_inclusive(heap_start_page, heap_end_page)
